Updated core code to support latest version of Hail on AWS
CloudFormation Template Changes:
* hail-master.template.yaml
    * Updated flow of Parameters
    * Added drop-downs for “Existing VPC ID” and “Existing Subnet ID”
    * Made Environment Type tag drop down optional
* hail-core.template.yaml
    * Updated lambda_handler to copy new emr-bootstrap-scripts folder that includes a shell script called ‘numpyinstall.sh’ that fixes the bug where EMR overwrites numpy with an older version of the package and breaks Hail.
* hail-ami.template.yaml
    * Changed ‘BuildSpec:’ to "./packer-files/buildspec.yml" in order to match updated packer-files folder in Quickstart.
    * Added additional Actions to ‘PackerMinimumBuildAccess’ IAM policy to allow Packer to pass role to EC2 instance (required).
    * Added ’s3:ListBucket’ action to ‘CodeBuildServiceAccess’ IAM policy.
    * Removed ‘hail-master’ and ‘hail-master_vep’ build projects due to Hail versioning scheme.  Could potentially bring back later.
* hail-emr.template.yaml
    * Added EMR Managed scaling policy to EMR cluster.

Packer Files Changes:
* Renamed ‘packer’ folder to ‘packer-files’.  Error was occurring with commands where Packer was confusing ‘packer’ folder and packer.exe.
* Updated build-wrapper.sh to pull from main version of Hail repository instead of depreciated master version for those running CodePipeline that pulls latest version of Hail with no VEP.
* Updated default version of EMR to ‘6.5.0’ and Spark to ‘3.1.2’ if end user does not specify these values in build-wrapper.sh.
* Updated packages to install/update in ‘install_prereqs’ function of hail_build.sh.
* Added additional conditions to ‘if’ logic statements that broke previous logic after Hail v 0.2.99 in hail_build.sh.
* Moved ‘buildspec.yml’ out of ‘packer-files’ folder and into ‘ami’ folder as a workaround for a CodeBuild bug when trying to find buildspec file.

General Changes:
* Added support for permission boundaries as well as custom prefixes and/or suffixes for all created roles and policies.

+++ packer-files/scripts/vep_install.sh
@@ -32,7 +32,8 @@ function install_prereqs {
         perl-Env \
         unzip \
         which \
-        zlib-devel
+        zlib-devel \
+        mariadb-devel
 
     cpanm \
         autodie \
@@ -53,8 +54,13 @@ function install_prereqs {
 
 # gsutil used to pull VEP 85 cache from the Broad
 function gsutil_install {
-  curl "$GSUTIL_SOURCE" | tar --directory "$GSUTIL_TARGET_DIR" -zx
-  echo "export PATH=\$PATH:$GSUTIL_TARGET_DIR/gsutil/" >> "$GSUTIL_PROFILE"
+    curl "$GSUTIL_SOURCE" | tar --directory "$GSUTIL_TARGET_DIR" -zx
+    echo "export PATH=\$PATH:$GSUTIL_TARGET_DIR/gsutil/" >> "$GSUTIL_PROFILE"
+}
+
+function download_homo_sapiens_file {
+    wget ftp://ftp.ensembl.org/pub/release-${1}/variation/vep/homo_sapiens_vep_${1}_${2}.tar.gz -P /tmp
+    # aws s3 cp /tmp/homo_sapiens_vep_${1}_${2}.tar.gz ${VEP_S3_SOURCE}${VEP_S3_CACHE_PATH}/
 }
 
 function vep_install {
@@ -75,24 +81,33 @@ function vep_install {
         HUMAN_REFERENCES=( GRCh37 GRCh38 )
         for REFERENCE in "${HUMAN_REFERENCES[@]}"
         do
+            if [ ! -f "/tmp/homo_sapiens_vep_${VEP_VERSION}_${REFERENCE}.tar.gz" ]; then
+                echo "/tmp/homo_sapiens_vep_${VEP_VERSION}_${REFERENCE}.tar.gz does not exist"
+                download_homo_sapiens_file ${VEP_VERSION} ${REFERENCE}
+            fi
+
             # Auto install (c)ache, and (f)asta
             tar --directory "$VEP_CACHE_DIR"  -xf "/tmp/homo_sapiens_vep_${VEP_VERSION}_$REFERENCE.tar.gz"
             perl INSTALL.pl --DESTDIR "$VEP_DIR" --CACHEDIR "$VEP_DIR"/cache --CACHEURL "$VEP_CACHE_DIR" \
-                 --AUTO cf --SPECIES homo_sapiens --ASSEMBLY "$REFERENCE" --NO_HTSLIB --NO_UPDATE
+                    --AUTO cf --SPECIES homo_sapiens --ASSEMBLY "$REFERENCE" --NO_HTSLIB --NO_UPDATE
             rm "/tmp/homo_sapiens_vep_${VEP_VERSION}_$REFERENCE.tar.gz"
         done
 
         # Rat (c)ache and (f)asta
-        tar --directory "$VEP_CACHE_DIR"  -xf "/tmp/rattus_norvegicus_vep_${VEP_VERSION}_Rnor_6.0.tar.gz"
-        perl INSTALL.pl --DESTDIR "$VEP_DIR" --CACHEDIR "$VEP_DIR"/cache --CACHEURL "$VEP_CACHE_DIR" \
-             --AUTO cf --SPECIES rattus_norvegicus --ASSEMBLY Rnor_6.0 --NO_HTSLIB --NO_UPDATE
-        rm "/tmp/rattus_norvegicus_vep_${VEP_VERSION}_Rnor_6.0.tar.gz"
+        if [ -f "/tmp/rattus_norvegicus_vep_${VEP_VERSION}_Rnor_6.0.tar.gz" ]; then
+            tar --directory "$VEP_CACHE_DIR"  -xf "/tmp/rattus_norvegicus_vep_${VEP_VERSION}_Rnor_6.0.tar.gz"
+            perl INSTALL.pl --DESTDIR "$VEP_DIR" --CACHEDIR "$VEP_DIR"/cache --CACHEURL "$VEP_CACHE_DIR" \
+                    --AUTO cf --SPECIES rattus_norvegicus --ASSEMBLY Rnor_6.0 --NO_HTSLIB --NO_UPDATE
+            rm "/tmp/rattus_norvegicus_vep_${VEP_VERSION}_Rnor_6.0.tar.gz"
+        fi
 
         # Zebrafish (c)ache and (f)asta
-        tar --directory "$VEP_CACHE_DIR"  -xf "/tmp/danio_rerio_vep_${VEP_VERSION}_GRCz11.tar.gz"
-        perl INSTALL.pl --DESTDIR "$VEP_DIR" --CACHEDIR "$VEP_DIR"/cache --CACHEURL "$VEP_CACHE_DIR" \
-             --AUTO cf --SPECIES danio_rerio --ASSEMBLY GRCz11 --NO_HTSLIB --NO_UPDATE
-        rm "/tmp/danio_rerio_vep_${VEP_VERSION}_GRCz11.tar.gz"
+        if [ -f "/tmp/danio_rerio_vep_${VEP_VERSION}_GRCz11.tar.gz" ]; then
+            tar --directory "$VEP_CACHE_DIR"  -xf "/tmp/danio_rerio_vep_${VEP_VERSION}_GRCz11.tar.gz"
+            perl INSTALL.pl --DESTDIR "$VEP_DIR" --CACHEDIR "$VEP_DIR"/cache --CACHEURL "$VEP_CACHE_DIR" \
+                    --AUTO cf --SPECIES danio_rerio --ASSEMBLY GRCz11 --NO_HTSLIB --NO_UPDATE
+            rm "/tmp/danio_rerio_vep_${VEP_VERSION}_GRCz11.tar.gz"
+        fi
 
         # Plugins are installed to $HOME.  Install all (p)lugins, then move to common location
         perl INSTALL.pl --AUTO p --PLUGINS all --NO_UPDATE
@@ -110,7 +125,23 @@ function vep_install {
     # Loftee for VEP GRCh37 only
     mkdir -p "$VEP_DIR"/loftee_data
     aws s3 sync "$VEP_S3_SOURCE$VEP_S3_LOFTEE_PATH" "$VEP_DIR"/loftee_data
+    if [ ! -f "${VEP_DIR}/loftee_data/phylocsf_gerp.sql.gz" ]; then
+        echo ""${VEP_DIR}"/loftee_data/phylocsf_gerp.sql.gz does not exist"
+        wget https://personal.broadinstitute.org/konradk/loftee_data/GRCh37/phylocsf_gerp.sql.gz -P ${VEP_DIR}/loftee_data
+        aws s3 cp ${VEP_DIR}/loftee_data/phylocsf_gerp.sql.gz "${VEP_S3_SOURCE}${VEP_S3_LOFTEE_PATH}/"
+    fi
     gunzip "$VEP_DIR"/loftee_data/phylocsf_gerp.sql.gz
+
+    # # Update below once Loftee is fixed for GRCh38
+    # # Loftee for VEP GRCh38
+    # mkdir -p "$VEP_DIR"/loftee_data/h38/
+    # wget https://personal.broadinstitute.org/konradk/loftee_data/GRCh38/loftee.sql.gz -P ${VEP_DIR}/loftee_data/h38
+    # wget https://personal.broadinstitute.org/konradk/loftee_data/GRCh38/gerp_conservation_scores.homo_sapiens.GRCh38.bw -P ${VEP_DIR}/loftee_data/h38
+    # wget https://personal.broadinstitute.org/konradk/loftee_data/GRCh38/human_ancestor.fa.gz -P ${VEP_DIR}/loftee_data/h38
+    # wget https://personal.broadinstitute.org/konradk/loftee_data/GRCh38/human_ancestor.fa.gz.fai -P ${VEP_DIR}/loftee_data/h38
+    # wget https://personal.broadinstitute.org/konradk/loftee_data/GRCh38/human_ancestor.fa.gz.gzi -P ${VEP_DIR}/loftee_data/h38
+    # # aws s3 sync "$VEP_DIR"/loftee_data/h38 "$VEP_S3_SOURCE$VEP_S3_LOFTEE_PATH/h38/"
+    # gunzip "$VEP_DIR"/loftee_data/h38/loftee.sql.gz
 }
 
 if [ "$VEP_VERSION" != "none" ]; then
